Se reorganizo el componente auth

+++ src/auth/modules/usuario/usuario.service.ts
@@ -1,58 +1,60 @@
 import { HttpException, Injectable, UnauthorizedException, NotFoundException } from '@nestjs/common';
 import { LoginDTO } from 'src/interfaces/login.dto';
 import { RegisterDTO } from 'src/interfaces/register.dto';
-import { IUsuario } from 'src/interfaces/user.interface';
-import { IPermiso } from 'src/interfaces/permisos.interface';
-import { UserEntity } from './user.entity';
+import { IUsuario } from 'src/auth/interfaces/user.interface';
+import { Usuario } from './usuario.entity';
 import { hashSync, compareSync } from 'bcrypt';
-import { JwtService } from 'src/jwt/jwt.service';
-import { Repository, DeepPartial} from 'typeorm';
-import { PermissionsService } from 'src/resources/permissions/permissions.service';
-import { RolesService } from 'src/resources/roles/roles.service';
-
+import { JwtService } from '../jwt/jwt.service';
+import { DeepPartial } from 'typeorm';
+import { PermisoService } from '../permiso/permiso.service';
+import { RolService } from '../rol/rol.service';
 
 @Injectable()
-export class UsersService {
-  repository = UserEntity; 
+export class UsuarioService {
+  repository = Usuario;
+
   constructor(
-    private permissionsService: PermissionsService,
-    private jwtService: JwtService, // Inyecta JwtService
-    private rolesService: RolesService
+    private permissionsService: PermisoService,
+    private jwtService: JwtService,
+    private rolesService: RolService
   ) {}
 
-  async createUsers(users: DeepPartial<UserEntity>) {
+  async createUsers(users: DeepPartial<Usuario>) {
     try {
       return await this.repository.save(users);
-  } catch (error) {
-      throw new HttpException('Create product error', 500)
-  }
-  }
-
-  async findUsers(): Promise<UserEntity[]>{
-      try {            
-        return await this.repository.find();
     } catch (error) {
-        throw new HttpException('Find all products error', 500)
-    } 
+      throw new HttpException('Create user error', 500);
+    }
   }
 
+  async findUsers(): Promise<Usuario[]> {
+    try {
+      return await this.repository.find();
+    } catch (error) {
+      throw new HttpException('Find all users error', 500);
+    }
+  }
 
-
-  async updateUserById(id: number, user: DeepPartial<UserEntity>): Promise<UserEntity> {
+  async updateUserById(id: number, user: DeepPartial<Usuario>): Promise<Usuario> {
     const query = this.repository.createQueryBuilder('user')
         .where('user.id = :id', { id });
     const userActual = await query.getOne();
-    this.repository.merge(userActual, user);
+
     if (!userActual) {
-        throw new NotFoundException(`User with id ${id} not found`);
+      throw new NotFoundException(`User with id ${id} not found`);
     }
+
+    this.repository.merge(userActual, user);
     return await this.repository.save(userActual);
   }
 
-  async deleteUserById(id: number): Promise <UserEntity> {
-    const userRemove = await this.repository.findOneBy({
-        id: id
-    })
+  async deleteUserById(id: number): Promise<Usuario> {
+    const userRemove = await this.repository.findOneBy({ id });
+
+    if (!userRemove) {
+      throw new NotFoundException(`User with id ${id} not found`);
+    }
+
     return await this.repository.remove(userRemove);
   }
 
@@ -64,7 +66,7 @@ export class UsersService {
     const hasPermission = usuario.roles.some(role => 
       role.permisos.some(permiso => permiso.nombre === nombrePermiso)
     );
-    
+
     if (!hasPermission) {
       throw new HttpException('El usuario no tiene el Permiso', 401);
     }
@@ -74,52 +76,50 @@ export class UsersService {
 
   async register(body: RegisterDTO) {
     try {
-      const user = new UserEntity();
+      const user = new Usuario();
       Object.assign(user, body);
       user.contrasena = hashSync(user.contrasena, 10);
       await this.repository.save(user);
-      return { status: 'created'};
+      return { status: 'created' };
     } catch (error) {
-      throw new HttpException('Error de creacion',500);
+      throw new HttpException('Error de creación', 500);
     }
   }
 
   async login(body: LoginDTO) {
     const user = await this.findByEmail(body.email);
-    if (user == null) {
-      throw new UnauthorizedException();
+
+    if (!user) {
+      throw new UnauthorizedException('User not found');
     }
 
     const compareResult = compareSync(body.password, user.contrasena);
+
     if (!compareResult) {
-      throw new UnauthorizedException();
+      throw new UnauthorizedException('Invalid password');
     }
 
     return {
       accessToken: this.jwtService.generateToken({ email: user.email }),
-      refreshToken: this.jwtService.generateToken(
-        { email: user.email },
-        'refresh',
-      ),
+      refreshToken: this.jwtService.generateToken({ email: user.email }, 'refresh'),
     };
   }
 
-  async findByEmail(email: string): Promise<UserEntity> {
+  async findByEmail(email: string): Promise<Usuario> {
     return await this.repository.findOneBy({ email });
   }
 
-
-  async assignPermissionToUser(idUsuario: number, body: { idPermiso: number }): Promise<UserEntity> {
-    
+  async assignPermissionToUser(idUsuario: number, body: { idPermiso: number }): Promise<Usuario> {
     const user = await this.repository.findOne({
       where: { id: idUsuario },
-      relations: ['permissions'],
+      relations: ['permisos'],
     });
 
     if (!user) {
       throw new NotFoundException(`User with ID ${idUsuario} not found`);
     }
-    const permission = await this.permissionsService.findPermissionById(body.idPermiso );
+
+    const permission = await this.permissionsService.findPermissionById(body.idPermiso);
 
     if (!permission) {
       throw new NotFoundException(`Permission with ID ${body.idPermiso} not found`);
@@ -128,14 +128,14 @@ export class UsersService {
     if (!user.permisos) {
       user.permisos = [];
     }
+
     user.permisos.push(permission); 
     await user.save();
     
     return user;
   }
 
-  async assignRoleToUser(userId: number, body: { roleId: number }): Promise<UserEntity> {
-    
+  async assignRoleToUser(userId: number, body: { roleId: number }): Promise<Usuario> {
     const user = await this.repository.findOne({
       where: { id: userId },
       relations: ['roles'],
@@ -144,6 +144,7 @@ export class UsersService {
     if (!user) {
       throw new NotFoundException(`User with ID ${userId} not found`);
     }
+
     const role = await this.rolesService.findRoleById(body.roleId);
 
     if (!role) {
@@ -153,6 +154,7 @@ export class UsersService {
     if (!user.roles) {
       user.roles = [];
     }
+
     user.roles.push(role);
     await user.save();
